util: relocate displayDetails.
