+++ lib/hsd-ledger-browser.js
@@ -6,6 +6,7 @@
 
 'use strict';
 
+const util = require('./utils/util');
 const USB = require('./device/usb');
 const WebAuthn = require('./device/webauthn');
 const DeviceError = require('./device/error');
@@ -15,6 +16,7 @@ const LedgerInput = require('./ledger/input');
 const LedgerChange = require('./ledger/change');
 const LedgerCovenant = require('./ledger/covenant');
 
+exports.util = util;
 exports.USB = USB;
 exports.WebAuthn = WebAuthn;
 exports.DeviceError = DeviceError;
