+++ lib/hsd-ledger.js
@@ -6,6 +6,7 @@
 
 'use strict';
 
+const util = require('./utils/util');
 const USB = require('./device/usb');
 const DeviceError = require('./device/error');
 const LedgerError = require('./ledger/error');
@@ -14,6 +15,7 @@ const LedgerInput = require('./ledger/input');
 const LedgerChange = require('./ledger/change');
 const LedgerCovenant = require('./ledger/covenant');
 
+exports.util = util;
 exports.USB = USB;
 exports.DeviceError = DeviceError;
 exports.LedgerError = LedgerError;
