+++ lib/utils/util.js
@@ -1,7 +1,8 @@
 'use strict';
 
+const assert = require('bsert');
 const bufio = require('bufio');
-const common = require('hsd/lib/hd/common');
+const {Address, hd, Rules} = require('hsd');
 
 /**
  * Index at which hardening begins.
@@ -9,7 +10,7 @@ const common = require('hsd/lib/hd/common');
  * @default
  */
 
-exports.HARDENED = common.HARDENED;
+exports.HARDENED = hd.common.HARDENED;
 
 /**
  * Return a hardened representation of the BIP44 index.
@@ -101,4 +102,60 @@ exports.splitBuffer = (data, size, zeroCopy = false) => {
  * @returns {Number[]}
  */
 
-exports.parsePath = common.parsePath;
+exports.parsePath = hd.common.parsePath;
+
+/**
+ * Displays mtx details.
+ * @param {Logger} logger
+ * @param {MTX} mtx
+ * @param {Object} options
+ */
+
+exports.displayDetails = (logger, network, mtx, options) => {
+  let fees = 0;
+
+  for (let i = 0; i < mtx.inputs.length; i++) {
+    const input = mtx.inputs[i];
+    const coin = mtx.view.getCoinFor(input);
+    fees += coin.value;
+  }
+
+  logger.info('Verify tx details on Ledger device.');
+  logger.info('');
+
+  for (let i = 0, j = 1; i < mtx.outputs.length; i++) {
+    const output = mtx.outputs[i];
+    fees -= output.value;
+
+    if (options && options.change && options.change.getIndex() === i)
+      continue;
+
+    logger.info(`Output #${j++}`);
+    logger.info(`Covenant: ${Rules.typesByVal[output.covenant.type]}`);
+
+    if (output.covenant.type !== Rules.types.NONE) {
+      assert(options && options.covenants, 'LedgerCovenants required.');
+
+      let name;
+
+      for (const covenant of options.covenants)
+        if (covenant.getIndex() === i)
+          name = covenant.getName();
+
+      logger.info(`Name: ${name}`);
+    }
+
+    if (output.covenant.type === Rules.types.TRANSFER) {
+      const ver = output.covenant.getU8(2);
+      const hash = output.covenant.get(3);
+      const addr = Address.fromHash(hash, ver);
+      logger.info(`New Owner: ${addr.toString(network)}`);
+    }
+
+    logger.info(`Value: ${output.value/1e6}`);
+    logger.info(`Address: ${output.address.toString('regtest')}`);
+    logger.info('');
+  }
+
+  logger.info(`Fees: ${fees/1e6}`);
+};
